Feature/ops cg cpp (#213)
* Initial files

* Added initial argparser and added Target class

* Added minor changes to argparser

* Initial commit of language.py, ops.py and store.py

* Added initial commit of c++ parser

* Jinja tests and filters created for custom environment and scheme class added

* cpp translator and scheme add

* Completed __main__.py. Need to add templates and schemes

* Initial end to end minor-bugfixes

* Fixing parsing arguments of templated function ops_par_loop and ops_decl_const

* Generic build fixes for ops_translator_v2

* mpi_seq templates added

* Makefile fix and verbosity improvements

* Initial correction of jinja templates

* Fixing templates and codegen errors

* code generation templates for seq completed with E2E compilation

* Jina utilities

* ops_arg_dat_opt support and user_type support added

* ops_arg_abl support bugfix

* Minor change in parser

* Feature/ops cg fortran (#197)

* fortran parsing

* rebased with features/ops-cg

* fixed few bugs with CPP parser and templates

* fixed bugs for CPP parser, able to generate and run seq,tiled,openmp,mpi,mpi_tiled and mpi_openmp versions

* tested TeaLeaf with new code-gen

* rebased with ops-cg

---------


* rebased with current develop branch

* initial commit for cuda cpp jinja template

* test commit to cuda template stage

* finished cuda/cpp template, tested with laplace, poisson, cloverleaf_3d and tealeaf application

* finished hip/cpp template, tested with laplace, poisson, cloverleaf_3d and tealeaf application

* tested most of the application from C application test-suite for following targets - seq tiled openmp mpi mpi_tiled mpi_openmp cuda mpi_cuda mpi_cuda_tiled, most of them are PASSED

* tested most of the application for OpenMP offload target with clang16.x

* corrected offset calculation for multidim dats when OPS_SOA not defined in templates

* added instructions for setting up python venv, added lattice-boltzman tutorial

* Update installation.md

Virtual Env installation instructions updated

* updated python instructions

* Tested apps with SYCL template added

* OpenACC template added

* Tested OpenMP offload MPI runs. Implemented workaround for too many arguments error for OpenMP_offload loop

* Adding new test scripts and env scripts

* added IEEE flags for clang

* Updated test scripts for Poisson

* Selective testing on Telos, AMOS and DMOS for Poisson App

* Update test.sh scripts

* Testing TeaLeaf with test scripts

* changed old code-gen to legacy and new code-gen as defualt

* Removing inline targets testing

* tested application which needs trdiagonal solver with new code-gen

* changed tridiagonal solver stragtegy for adi

* tested sycl targets with new and legacy code gen

* changed stride calclulation in sycl template

* added mpi_openmp target for Fortran

* corrected create_kerneldesc function call

* libclang added to requirments of new translator (#212)

* libclang added to requirments of new translator

* Added llvmlite pip requirement

* Fixing version issue in pkgs

* minor correction

* Update test script for CloverLeaf and source files

* Updating test.sh scrpts

* Further changes to test scripts

---------


* avoided printing 'stddef.h' file not found diagnostic message

* Updaes to test.sh scripts for smaller applications

* added prints for code-gen progress happening

* Initial files

* Added initial argparser and added Target class

* Added minor changes to argparser

* Initial commit of language.py, ops.py and store.py

* Added initial commit of c++ parser

* Jinja tests and filters created for custom environment and scheme class added

* cpp translator and scheme add

* Completed __main__.py. Need to add templates and schemes

* Initial end to end minor-bugfixes

* Fixing parsing arguments of templated function ops_par_loop and ops_decl_const

* Generic build fixes for ops_translator_v2

* mpi_seq templates added

* Makefile fix and verbosity improvements

* Initial correction of jinja templates

* Fixing templates and codegen errors

* code generation templates for seq completed with E2E compilation

* Jina utilities

* ops_arg_dat_opt support and user_type support added

* ops_arg_abl support bugfix

* Minor change in parser

* Feature/ops cg fortran (#197)

* fortran parsing

* rebased with features/ops-cg

* fixed few bugs with CPP parser and templates

* fixed bugs for CPP parser, able to generate and run seq,tiled,openmp,mpi,mpi_tiled and mpi_openmp versions

* tested TeaLeaf with new code-gen

* rebased with ops-cg

---------


* rebased with current develop branch

* initial commit for cuda cpp jinja template

* test commit to cuda template stage

* finished cuda/cpp template, tested with laplace, poisson, cloverleaf_3d and tealeaf application

* finished hip/cpp template, tested with laplace, poisson, cloverleaf_3d and tealeaf application

* tested most of the application from C application test-suite for following targets - seq tiled openmp mpi mpi_tiled mpi_openmp cuda mpi_cuda mpi_cuda_tiled, most of them are PASSED

* tested most of the application for OpenMP offload target with clang16.x

* corrected offset calculation for multidim dats when OPS_SOA not defined in templates

* added instructions for setting up python venv, added lattice-boltzman tutorial

* Update installation.md

Virtual Env installation instructions updated

* updated python instructions

* Tested apps with SYCL template added

* OpenACC template added

* Tested OpenMP offload MPI runs. Implemented workaround for too many arguments error for OpenMP_offload loop

* Adding new test scripts and env scripts

* added IEEE flags for clang

* Updated test scripts for Poisson

* Selective testing on Telos, AMOS and DMOS for Poisson App

* Update test.sh scripts

* Testing TeaLeaf with test scripts

* changed old code-gen to legacy and new code-gen as defualt

* Removing inline targets testing

* tested application which needs trdiagonal solver with new code-gen

* changed tridiagonal solver stragtegy for adi

* tested sycl targets with new and legacy code gen

* changed stride calclulation in sycl template

* added mpi_openmp target for Fortran

* corrected create_kerneldesc function call

* libclang added to requirments of new translator (#212)

* libclang added to requirments of new translator

* Added llvmlite pip requirement

* Fixing version issue in pkgs

* minor correction

* Update test script for CloverLeaf and source files

* Updating test.sh scrpts

* Further changes to test scripts

---------


* avoided printing 'stddef.h' file not found diagnostic message

* added prints for code-gen progress happening

* fixed problems after rebase

* fixed problem with cloverleaf_3d_hdf5 sequantial backend

* fixed problem due to adding half precison header files

* Updating test scripts and minor fixes after merge

* updated as per review comment

* updated ops_compiler check with wildcard function

---------

+++ apps/c/multiDim/test.sh
@@ -1,15 +1,26 @@
 #!/bin/bash
 set -e
 cd $OPS_INSTALL_PATH/c
-#<<COMMENT
-if [[ -v HIP_INSTALL_PATH ]]; then
-  
-  echo '============> NO Test for multiDim HIP'
 
-fi
+export SOURCE_INTEL=source_intel_2021.3_pythonenv
+export SOURCE_PGI=source_pgi_nvhpc_23_pythonenv
+export SOURCE_INTEL_SYCL=source_intel_2021.3_sycl_pythonenv
+export SOURCE_AMD_HIP=source_amd_rocm-5.4.3_pythonenv
 
+export AMOS=TRUE
+#export DMOS=TRUE
+#export TELOS=TRUE
+#export KOS=TRUE
+
+if [[ -v TELOS || -v KOS ]]; then
+
+#============================ Test with Intel Classic Compilers==========================================
+echo "Testing Intel classic complier based applications ---- "  
+cd $OPS_INSTALL_PATH/c
 source ../../scripts/$SOURCE_INTEL
-make -j
+
+make clean
+make
 cd $OPS_INSTALL_PATH/../apps/c/multiDim/
 
 make clean
@@ -17,7 +28,6 @@ rm -f .generated
 make IEEE=1 -j
 
 
-#============================ Test multidim with Intel Compilers ==========================================
 echo '============> Running OpenMP'
 KMP_AFFINITY=compact OMP_NUM_THREADS=1,12 ./multidim_openmp > perf_out
 grep "Reduction result" perf_out
@@ -59,7 +69,7 @@ grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm perf_out
 
-
+if [[ -v CUDA_INSTALL_PATH ]]; then
 echo '============> Running CUDA'
 ./multidim_cuda OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 grep "Reduction result" perf_out
@@ -84,54 +94,70 @@ rm perf_out
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 #rm perf_out
 
-echo '============> Running OpenCL on CPU'
-OMP_NUM_THREADS=1 ./multidim_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=512 OPS_BLOCK_SIZE_Y=1 > perf_out
+fi
+
+echo "All Intel classic complier based applications ---- PASSED"
+fi
+
+if [[ -v TELOS ]]; then
+
+echo "Testing Intel SYCL complier based applications ---- "
+
+cd $OPS_INSTALL_PATH/c
+source ../../scripts/$SOURCE_INTEL_SYCL
+#make -j -B
+make clean
+make
+cd $OPS_INSTALL_PATH/../apps/c/multiDim/
+
+make clean
+#make IEEE=1 -j
+make IEEE=1 multidim_sycl multidim_mpi_sycl multidim_mpi_sycl_tiled
+
+if [[ -v CUDA_INSTALL_PATH ]]; then
+echo '============> Running SYCL'
+./multidim_sycl OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 grep "Reduction result" perf_out
 grep "Total Wall time" perf_out
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm perf_out
 
-echo '============> Running OpenCL on GPU'
-OMP_NUM_THREADS=1 ./multidim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-OMP_NUM_THREADS=1 ./multidim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
+echo '============> Running MPI+SYCL'
+OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_sycl OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 grep "Reduction result" perf_out
 grep "Total Wall time" perf_out
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm perf_out
 
-echo '============> Running MPI+OpenCL on CPU'
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 4 ./multidim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 4 ./multidim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
+echo '============> Running MPI+SYCL Tiled'
+OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_sycl_tiled OPS_TILING OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 grep "Reduction result" perf_out
 grep "Total Wall time" perf_out
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm perf_out
 
-echo '============> Running MPI+OpenCL on GPU'
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-grep "Reduction result" perf_out
-grep "Total Wall time" perf_out
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm perf_out
+echo "All Intel SYCL complier based applications ---- PASSED"
 
-echo "All Intel complied applications PASSED : Moving no to PGI Compiler Tests "
+fi
 
 
-cd -
+if [[ -v TELOS ]]; then
+
+#============================ Test with PGI Compilers==========================================
+echo "Testing PGI/NVHPC complier based applications ---- "
+
+cd $OPS_INSTALL_PATH/c
 source ../../scripts/$SOURCE_PGI
 
 make clean
 make 
-cd -
+cd $OPS_INSTALL_PATH/../apps/c/multiDim/
 make clean
 make
 
-#============================ Test multidim with PGI Compilers ==========================================
 echo '============> Running OpenMP'
 KMP_AFFINITY=compact OMP_NUM_THREADS=1,12 ./multidim_openmp > perf_out
 grep "Reduction result" perf_out
@@ -164,6 +190,7 @@ grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm perf_out
 
+if [[ -v CUDA_INSTALL_PATH ]]; then
 echo '============> Running CUDA'
 ./multidim_cuda OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 grep "Reduction result" perf_out
@@ -188,55 +215,81 @@ rm perf_out
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 #rm perf_out
 
-#echo '============> Running OpenCL on CPU'
-#./multidim_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=512 OPS_BLOCK_SIZE_Y=1 > perf_out
+#echo '============> Running OpenACC'
+#./multidim_openacc OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
 #grep "Reduction result" perf_out
 #grep "Total Wall time" perf_out
 #grep "PASSED" perf_out
 #rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 #rm perf_out
-
-echo '============> Running OpenCL on GPU'
-OMP_NUM_THREADS=1 ./multidim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-OMP_NUM_THREADS=1 ./multidim_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-grep "Reduction result" perf_out
-grep "Total Wall time" perf_out
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm perf_out
-
-#echo '============> Running MPI+OpenCL on CPU'
-#$MPI_INSTALL_PATH/bin/mpirun -np 20 ./multidim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
-#$MPI_INSTALL_PATH/bin/mpirun -np 20 ./multidim_mpi_opencl OPS_CL_DEVICE=0 OPS_BLOCK_SIZE_X=256 OPS_BLOCK_SIZE_Y=1 > perf_out
-#grep "Reduction result" perf_out
-#grep "Total Wall time" perf_out
-#grep "PASSED" perf_out
-#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-#rm perf_out
-
-echo '============> Running MPI+OpenCL on GPU'
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_opencl OPS_CL_DEVICE=1 OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-grep "Reduction result" perf_out
-grep "Total Wall time" perf_out
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm perf_out
-
-echo '============> Running OpenACC'
-./multidim_openacc OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
-grep "Reduction result" perf_out
-grep "Total Wall time" perf_out
-grep "PASSED" perf_out
-rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
-rm perf_out
 #COMMENT
-echo '============> Running MPI+OpenACC'
-OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_openacc OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
+#echo '============> Running MPI+OpenACC'
+#OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_openacc OPS_BLOCK_SIZE_X=32 OPS_BLOCK_SIZE_Y=4 > perf_out
+#grep "Reduction result" perf_out
+#grep "Total Wall time" perf_out
+#grep "PASSED" perf_out
+#rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
+#rm perf_out
+#echo "All PGI complied applications PASSED : Exiting Test Script "
+
+fi
+
+echo '============> Running OMPOFFLOAD'
+./multidim_offload OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
 grep "Reduction result" perf_out
 grep "Total Wall time" perf_out
 grep "PASSED" perf_out
 rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
 rm perf_out
-echo "All PGI complied applications PASSED : Exiting Test Script "
 
+echo '============> Running MPI+OMPOFFLOAD'
+OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_offload OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
+grep "Reduction result" perf_out
+grep "Total Wall time" perf_out
+grep "PASSED" perf_out
+rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
+rm perf_out
+
+
+fi
+
+echo "All PGI complier based applications ---- PASSED"
+
+fi
+
+if [[ -v AMOS ]]; then
+
+echo "Testing AMD HIP complier based applications ---- "
+cd $OPS_INSTALL_PATH/c
+source ../../scripts/$SOURCE_AMD_HIP
+#make -j -B
+#make clean
+#make 
+cd $OPS_INSTALL_PATH/../apps/c/multiDim
+
+rm -f .generated
+make clean
+#make IEEE=1 -j
+make multidim_hip multidim_mpi_hip #multiDim_hip_tiled multiDim_mpi_hip_tiled
+
+echo '============> Running HIP'
+./multidim_hip OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
+grep "Reduction result" perf_out
+grep "Total Wall time" perf_out
+grep "PASSED" perf_out
+rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
+rm perf_out
+
+echo '============> Running MPI+HIP'
+OMP_NUM_THREADS=1 $MPI_INSTALL_PATH/bin/mpirun -np 2 ./multidim_mpi_hip OPS_BLOCK_SIZE_X=64 OPS_BLOCK_SIZE_Y=4 > perf_out
+grep "Reduction result" perf_out
+grep "Total Wall time" perf_out
+grep "PASSED" perf_out
+rc=$?; if [[ $rc != 0 ]]; then echo "TEST FAILED";exit $rc; fi
+rm perf_out
+
+echo "All AMD HIP complier based applications ---- PASSED"
+
+fi
+
+echo "---------- Exiting Test Script "
